refactor: reorganise project structure
