+++ src/Odotocodot.OneNote.Linq/Classes/Notebook.cs
