+++ src/Odotocodot.OneNote.Linq/Classes/OpenSections.cs
