+++ src/Odotocodot.OneNote.Linq/Classes/Section.cs
