+++ src/Odotocodot.OneNote.Linq/Classes/SectionGroup.cs
