+++ src/Odotocodot.OneNote.Linq/Enums/ComObjectMode.cs
