+++ src/Odotocodot.OneNote.Linq/Enums/OpenMode.cs
